first rest tests

+++ src/main/java/pl/coderstrust/accounting/controller/InvoiceController.java
@@ -1,7 +1,7 @@
 package pl.coderstrust.accounting.controller;
 
-import io.swagger.annotations.Api;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import pl.coderstrust.accounting.exceptions.InvoiceNotFoundException;
 import pl.coderstrust.accounting.logic.InvoiceService;
@@ -17,9 +16,8 @@ import pl.coderstrust.accounting.model.Invoice;
 
 import java.io.IOException;
 import java.util.Collection;
+import java.util.Optional;
 
-@Api(value = "/invoices", description = "Invoicing system")
-//@Produces({"application/json", "application/xml"})
 @RestController
 @RequestMapping("/invoices")
 public class InvoiceController {
@@ -38,23 +36,52 @@ public class InvoiceController {
     return invoiceService.getInvoices();
   }
 
+  @GetMapping("/{id}")
+  public ResponseEntity<?> getInvoice(@PathVariable int id) throws IOException {
+    Optional<Invoice> invoiceOptional = invoiceService.getInvoices().stream()
+        .filter(invoice -> invoice.getId().equals(id)).findAny();
+
+    if (!invoiceOptional.isPresent()) {
+      return ResponseEntity.notFound().build();
+    }
+
+    return ResponseEntity.ok(invoiceOptional.get());
+  }
+
   @PostMapping
   public int saveInvoice(@RequestBody Invoice invoice) throws IOException {
     return invoiceService.saveInvoice(invoice);
   }
 
   @PutMapping("/{id}")
-  public void updateInvoice(@RequestBody Invoice invoice) throws IOException, InvoiceNotFoundException {
-    invoiceService.updateInvoice(invoice.getId(), invoice);
+  public ResponseEntity<?> updateInvoice(@PathVariable int id, @RequestBody Invoice updatedInvoice)
+      throws IOException, InvoiceNotFoundException {
+
+    Optional<Invoice> invoiceOptional = invoiceService.getInvoices().stream()
+        .filter(invoice -> invoice.getId().equals(id)).findAny();
+
+    if (!invoiceOptional.isPresent()) {
+      return ResponseEntity.notFound().build();
+    }
+
+    invoiceService.updateInvoice(id, updatedInvoice);
+
+    return ResponseEntity.ok().build();
   }
 
   @DeleteMapping("/{id}")
-  public void removeInvoiceById(@PathVariable int id) throws InvoiceNotFoundException, IOException {
+  public ResponseEntity<?> removeInvoiceById(int id) throws InvoiceNotFoundException, IOException {
+
+    Optional<Invoice> invoiceOptional = invoiceService.getInvoices().stream()
+        .filter(invoice -> invoice.getId().equals(id)).findAny();
+
+    if (!invoiceOptional.isPresent()) {
+      return ResponseEntity.notFound().build();
+    }
+
     invoiceService.removeInvoiceById(id);
+
+    return ResponseEntity.ok().build();
   }
 
-  @GetMapping("/{id}")
-  public Invoice getInvoiceById(@PathVariable int id) throws IOException, InvoiceNotFoundException {
-    return invoiceService.getInvById(id);
-  }
 }
